+++ src/components/Loading.jsx
@@ -1,5 +1,11 @@
+import Logo from "../assets/peeko-logo.png";
+
 const Loading = () => {
-    return <div>Loading...</div>;
+    return (
+        <div className="bg-slate-800 min-h-svh flex justify-center items-center">
+            <img src={Logo} alt="Logo" className="loading-logo w-32" />
+        </div>
+    );
 };
 
 export default Loading;